+++ python/packages/core/agent_framework/_tools.py
@@ -2195,7 +2195,7 @@ class FunctionInvocationLayer(Generic[OptionsCoT]):
                 # Loop exhausted all iterations (or function invocation disabled).
                 # Make a final model call with tool_choice="none" so the model
                 # produces a plain text answer instead of leaving orphaned
-                # function_call items without matching results (issue #1366).
+                # function_call items without matching results.
                 if response is not None and self.function_invocation_configuration["enabled"]:
                     logger.info(
                         "Maximum iterations reached (%d). Requesting final response without tools.",
@@ -2334,7 +2334,7 @@ class FunctionInvocationLayer(Generic[OptionsCoT]):
             # Loop exhausted all iterations (or function invocation disabled).
             # Make a final model call with tool_choice="none" so the model
             # produces a plain text answer instead of leaving orphaned
-            # function_call items without matching results (issue #1366).
+            # function_call items without matching results.
             if response is not None and self.function_invocation_configuration["enabled"]:
                 logger.info(
                     "Maximum iterations reached (%d). Requesting final response without tools.",
